Add files via upload
Puzzles are mostly done. Mercenary does not work as he should; gambling puzzle doesn't have limits, nor does it work due to player gold not being updated; Sunstone doesn't cost anything (not sure where that code is); Rhyming puzzle doesn't give silver dagger and the dagger cannot be equipped nor used on the wraith (equipping and making it an instant kill on the wraith would probably be easier). I believe that is all.